+++ apps/sv/src/main/scala/org/lfdecentralizedtrust/splice/sv/automation/DsoDelegateBasedAutomationService.scala
@@ -20,6 +20,7 @@ import org.lfdecentralizedtrust.splice.store.{
 import org.lfdecentralizedtrust.splice.scan.admin.api.client.ScanConnection
 import org.lfdecentralizedtrust.splice.sv.automation.delegatebased.*
 import org.lfdecentralizedtrust.splice.sv.automation.delegatebased.ExpiredAmuletAllocationTrigger
+import org.lfdecentralizedtrust.splice.sv.store.IgnoredPartiesStore
 import org.lfdecentralizedtrust.splice.sv.config.SvAppBackendConfig
 
 import scala.concurrent.{ExecutionContextExecutor, Future}
@@ -49,6 +50,10 @@ class DsoDelegateBasedAutomationService(
       : org.lfdecentralizedtrust.splice.sv.automation.DsoDelegateBasedAutomationService.type =
     DsoDelegateBasedAutomationService
 
+  val expiredAmuletIgnoredPartiesStore = new IgnoredPartiesStore(
+    triggerContext.config.ignoredExpiredAmuletPartyIds
+  )
+
   def start(): Unit = {
     registerTrigger(new AdvanceOpenMiningRoundTrigger(triggerContext, svTaskContext))
     registerTrigger(new UpdateExternalPartyConfigStateTrigger(triggerContext, svTaskContext))
@@ -63,13 +68,37 @@ class DsoDelegateBasedAutomationService(
     }
     registerTrigger(new MergeMemberTrafficContractsTrigger(triggerContext, svTaskContext))
 
-    registerTrigger(new ExpiredAmuletTrigger(config, triggerContext, svTaskContext))
-    registerTrigger(new ExpiredLockedAmuletTrigger(config, triggerContext, svTaskContext))
     registerTrigger(
-      new ExpiredAmuletTransferInstructionTrigger(config, clock, triggerContext, svTaskContext)
+      new ExpiredAmuletTrigger(
+        config,
+        triggerContext,
+        svTaskContext,
+        expiredAmuletIgnoredPartiesStore,
+      )
     )
     registerTrigger(
-      new ExpiredAmuletAllocationTrigger(config, clock, triggerContext, svTaskContext)
+      new ExpiredLockedAmuletTrigger(
+        config,
+        triggerContext,
+        svTaskContext,
+        expiredAmuletIgnoredPartiesStore,
+      )
+    )
+    registerTrigger(
+      new ExpiredAmuletTransferInstructionTrigger(
+        config,
+        clock,
+        triggerContext,
+        svTaskContext,
+      )
+    )
+    registerTrigger(
+      new ExpiredAmuletAllocationTrigger(
+        config,
+        clock,
+        triggerContext,
+        svTaskContext,
+      )
     )
     registerTrigger(new ExpiredSvOnboardingRequestTrigger(triggerContext, svTaskContext))
     registerTrigger(new CloseVoteRequestTrigger(triggerContext, svTaskContext))
