+++ test-full-class-names.log
@@ -1,6 +1,7 @@
 org.lfdecentralizedtrust.splice.integration.tests.AmuletAllocationsIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.AmuletExpiryIntegrationTest
-org.lfdecentralizedtrust.splice.integration.tests.AmuletWithMinimalVersionExpiryIntegrationTest
+org.lfdecentralizedtrust.splice.integration.tests.AmuletExpiryWithBrokenPackageIntegrationTest
+org.lfdecentralizedtrust.splice.integration.tests.AmuletExpiryWithMinimalPackageIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.Ans4SvsIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.AnsIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.AutomationControlIntegrationTest
